(chore) Fixes address related tests because of mock and options API changes.
